# apps/backend/src/shared/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mandarin Learning Platform API
  version: 1.0.0
  description: |
    Production backend API for Mandarin vocabulary learning and AI-powered
    conversation generation. All endpoints are versioned under /api/v1.

    This reference documents the full NestJS 11 route surface: 54 route paths
    across the 15 backend modules (audio, auth, characters, chengyu,
    foundations, grammar, health, mnemonics, phonetic-clusters, progression,
    quiz, radicals, readers, review, words — pinyin lives under foundations).

    Authentication model (per route):
      - Public — no auth required (static reference data, health, auth
        register/login/refresh/logout).
      - Required auth — `Authorization: Bearer <access_token>`; user-scoped
        state or AI/vendor-cost generation (`RequireAuthGuard`).
      - Optional auth — bearer token is best-effort (`OptionalAuthGuard`): an
        absent or invalid token proceeds as a guest (`req.userId` undefined),
        never 401. Guest responses are session-local/empty shapes.

    All 4xx/5xx errors use the `{ code, message, requestId }` envelope.
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://mandarin-vite-react-ts-production.up.railway.app/api
    description: Production server (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HTTP-only refresh token cookie
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Standard error envelope — the global 24-3 `AppExceptionFilter`
        serializes every thrown `HttpException` into this shape.
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request ID (echoes the `X-Request-Id` response header)
      required:
        - code
        - message
        - requestId
...
paths:
...
  /v1/radicals/character/{glyph}:
    get:
      summary: Get radicals composing a character
      tags:
        - Radicals
      description: Returns radicals that compose the given character glyph. Public.
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
      responses:
        "200":
          description: Radicals composing the character
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "500":
          description: Failed to load radicals for character (LOAD_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/characters/{glyph}:
    get:
      summary: Get character by glyph
      tags:
        - Characters
      description: |
        Returns full character detail (pinyin, meanings, stroke count, radical,
        etc.). Public — no authentication.

        **Shadow note:** this path is served by the FOUNDATIONS module — the
        foundations router is mounted before the characters router, so its
        `characters/:glyph` handler captures every single-segment
        `GET /v1/characters/<x>` on the live route tree (shadowing the
        characters module's own `:glyph`, `/search`, and `/frequency`
        handlers). Unknown glyphs return 404 `Character "<glyph>" not found`.
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
      responses:
        "200":
          description: Character detail
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Missing glyph (VALIDATION_ERROR — HTTP-unreachable for single-segment paths)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Character not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to get character data (LOAD_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/characters/{glyph}/phonetic:
    get:
      summary: Get phonetic component
      tags:
        - Characters
      description: |
        Returns phonetic component info (glyph, pinyin, meaning). Public.
        Served by the characters module (not shadowed — multi-segment path).
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
      responses:
        "200":
          description: Phonetic component info
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid character glyph (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Character or phonetic component not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to get phonetic component (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/characters/{glyph}/homophones:
    get:
      summary: Get homophones
      tags:
        - Characters
      description: |
        Returns all characters sharing the same pronunciation, grouped by
        reading. Public.
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
        - name: exactTone
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to exact tone match
      responses:
        "200":
          description: Homophones grouped by reading
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid character glyph (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Character not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to get homophones (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/characters/{glyph}/decomposition:
    get:
      summary: Get character decomposition
      tags:
        - Characters
      description: |
        Returns the decomposition tree — constituent components with types.
        Public.
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
      responses:
        "200":
          description: Decomposition tree
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid character glyph (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Character not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to get decomposition (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/characters/search:
    get:
      summary: Search characters (SHADOWED — returns 404 in practice)
      tags:
        - Characters
      description: |
        Searches characters by pinyin, tone filter, or HSK level. Declared on
        the characters controller AFTER `:glyph`.

        **Shadow note:** on the live route tree the foundations `:glyph` shadow
        matches `search` first, so this path returns 404
        `Character "search" not found`. The underlying 2xx search logic is
        covered by service unit tests; the full route activates only if the
        route order is reordered.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Pinyin search query
        - name: tone
          in: query
          required: false
          schema:
            type: integer
          description: Tone filter
        - name: hskLevel
          in: query
          required: false
          schema:
            type: integer
          description: HSK level filter
      responses:
        "200":
          description: Search results (only when the shadow is lifted)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
        "404":
          description: Shadowed by the foundations :glyph route (practical behavior)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "400":
          description: Validation error (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to search characters (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/characters/frequency:
    get:
      summary: Get characters by frequency (SHADOWED — returns 404 in practice)
      tags:
        - Characters
      description: |
        Returns characters ordered by frequency rank, optionally filtered by
        HSK tier. Declared on the characters controller AFTER `:glyph`.

        **Shadow note:** on the live route tree the foundations `:glyph` shadow
        matches `frequency` first, so this path returns 404
        `Character "frequency" not found`. The underlying 2xx frequency logic
        is covered by service unit tests.
      parameters:
        - name: tier
          in: query
          required: false
          schema:
            type: integer
          description: HSK tier filter (1-6)
        - name: page
          in: query
          required: false
          schema:
            type: integer
          description: Page number (default 1)
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
          description: Page size (default 50)
      responses:
        "200":
          description: Frequency list (only when the shadow is lifted)
          content:
            application/json:
              schema:
                type: array
        "404":
          description: Shadowed by the foundations :glyph route (practical behavior)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "400":
          description: Invalid tier parameter (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to get frequency list (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ══════════════════════════════════════════════════════════════════
  # Words
  # ══════════════════════════════════════════════════════════════════
...